figo_fagot/requirements.txt
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
idna==2.8
Morfessor==2.0.6
numpy==1.17.3
pandas==0.25.2
polyglot==16.7.4
pycld2==0.41
PyICU==2.3.1
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
six==1.12.0
soupsieve==1.9.4
urllib3==1.25.6
